+++ theme2/adm/affiliation_update.php
@@ -42,6 +42,12 @@ else if ($_POST['act_button'] == "선택수정" && $w == 'u')
             // 실제 번호를 넘김
             $k = isset($_POST['chk'][$i]) ? (int) $_POST['chk'][$i] : 0;
 
+            // 회원데이터에 계열값이 들어있는지 확인
+            $overlab = sql_fetch("SELECT count(*) as cnt from {$g5['member_table']} where affiliation = '{$af_name[$k]}' AND reunion_id = $reunionID");
+            if($overlab['cnt'] > 0){
+                alert("회원 데이터가 있어 삭제 할 수 없습니다.");
+                return false;
+            }
             sql_query("DELETE FROM `affiliation` WHERE af_id = '{$_POST['af_id'][$k]}' AND reunion_id = '{$reunionID}'");
         }
